requirements.txt
alembic==1.11.3 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
greenlet==2.0.2 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
pytest==7.4.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.20 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
